adding javadocs, license, and author tags to newest classes; deleting unused TravelGraphSchema
