+++ .gitignore
@@ -18,4 +18,3 @@
 *~
 .*~
 !/src/test/java/com/amazon/janusgraph/diskstorage/dynamodb/
-rebuild-backend.sh
