+++ src/main/java/com/amazon/janusgraph/batch/BatchCommand.java
@@ -1,3 +1,17 @@
+/*
+ * Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
 package com.amazon.janusgraph.batch;
 
 import org.apache.commons.lang.exception.ExceptionUtils;
@@ -7,7 +21,9 @@ import org.slf4j.Logger;
 import java.util.concurrent.BlockingQueue;
 
 /**
- * Created by addisonslabaugh on 6/13/17.
+ *
+ * @author Addison Slabaugh
+ *
  */
 public class BatchCommand implements Runnable {
