+++ src/main/java/com/amazon/janusgraph/example/AbstractGraph.java
@@ -1,4 +1,18 @@
-package com.amazon.janusgraph.graph;
+/*
+ * Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package com.amazon.janusgraph.example;
 
 import com.amazon.janusgraph.triple.Triple;
 import org.janusgraph.core.JanusGraph;
@@ -11,7 +25,9 @@ import java.util.concurrent.*;
 
 
 /**
- * Created by addisonslabaugh on 6/21/17.
+ *
+ * @author Addison Slabaugh
+ *
  */
 public interface AbstractGraph {
