+++ src/main/java/com/amazon/janusgraph/example/TravelGraph.java
@@ -1,4 +1,18 @@
-package com.amazon.janusgraph.graph;
+/*
+ * Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package com.amazon.janusgraph.example;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -17,6 +31,11 @@ import com.codahale.metrics.MetricRegistry;
 
 import com.amazon.janusgraph.triple.Triple;
 
+/**
+ *
+ * @author Addison Slabaugh
+ *
+ */
 public class TravelGraph implements AbstractGraph {
 
     public static final MetricRegistry REGISTRY = MetricManager.INSTANCE.getRegistry();
@@ -24,6 +43,12 @@ public class TravelGraph implements AbstractGraph {
     private static final int POOL_SIZE = 10;
     private static Set<Triple> triples;
 
+    /**
+     *
+     * @param graph
+     * @param report        allows client to determine whether or not to generate a report after ingestion
+     * @throws Exception
+     */
     public void load(final JanusGraph graph, boolean report) throws Exception {
 
         URL resource = classLoader.getResource("META-INF/HotelTriples.txt");
@@ -49,6 +74,14 @@ public class TravelGraph implements AbstractGraph {
         LOG.info("TravelGraph.load complete");
     }
 
+    /**
+     * A method that processes a file containing left and right objects, properties, and relationships and
+     * returns a set of triples.
+     *
+     * @param url   URL that points to a resource containing left and right objects, properties
+     *              and relationships
+     * @return      a set of triples that can be ingested into a graph
+     */
     @Override
     public Set<Triple> processFile(URL url) {
         Set<Triple> triples = new HashSet<>();
