+++ src/test/java/com/amazon/janusgraph/MarvelTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertTrue;
 import java.util.Collection;
 import java.util.Iterator;
 
-import com.amazon.janusgraph.graph.MarvelGraph;
+import com.amazon.janusgraph.example.MarvelGraph;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource;
 import org.apache.tinkerpop.gremlin.structure.Direction;
 import org.apache.tinkerpop.gremlin.structure.Edge;
